chore(regen): Random.cs picks up the shift-count cast removal from 8aded6349 (#1304)
The emitter's dead 'int32' arm made it treat every int shift count as
non-int and insert a redundant (int) cast; 8aded6349 revived the arm but
the spy-sourced stdlib was not regenerated with it. Only Random.cs
changes (1 << (int)(k) -> 1 << k, a no-op cast removal); all 59 Random
tests pass and the staleness checker is clean again.

+++ src/Sharpy.Stdlib/Random/Random.cs
@@ -167,7 +167,7 @@ namespace Sharpy
                 throw new global::Sharpy.ValueError("number of bits must be <= 30 for int return");
             }
 
-            return _Random.Next(1 << (int)(k));
+            return _Random.Next(1 << k);
         }
 
         /// <summary>
